3496 - Refactor GeoOptions components hierarchy
