+++ src/client/components/PageLayout/Toolbar/Options/GeoOptions/Center/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Center'
