--- src/client/components/PageLayout/Toolbar/Options/GeoOptions/GeoOptions.scss
@@ -1,8 +1,0 @@
-@import 'src/client/style/partials';
-
-.geo-route-options__group {
-  align-items: center;
-  display: flex;
-  gap: $spacing-xxs;
-  height: 100%;
-}
