+++ src/client/components/PageLayout/Toolbar/Options/GeoOptions/GeoToggle/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GeoToggle'
