+++ src/client/components/PageLayout/Toolbar/Options/GeoOptions/MapType/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MapType'
