+++ src/client/components/PageLayout/Toolbar/Options/GeoOptions/Zoom/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Zoom'
