+++ src/client/pages/Geo/Map/Map.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 
 import { useGeoMap } from './hooks/useGeoMap'
 import { useMapLayersHandler } from './hooks/useMapLayersHandler'
+import { useMapOptionsUpdateListeners } from './hooks/useMapOptionsUpdateListeners'
 
 type Props = {
   viewport?: google.maps.LatLngBoundsLiteral
@@ -12,8 +13,8 @@ const Map: React.FC<React.PropsWithChildren<Props>> = (props) => {
   const { viewport, children } = props
 
   const { map, ref } = useGeoMap({ viewport })
-
   useMapLayersHandler()
+  useMapOptionsUpdateListeners()
 
   return (
     <>
